+++ src/targets/openbsd.rs
@@ -1,12 +1,9 @@
-use crate::config::{AppError, Config, FetchMirrors, LogFormatter};
+use crate::config::{fetch_text, AppError, Config, FetchMirrors, LogFormatter};
 use crate::countries::Country;
 use crate::mirror::Mirror;
 use crate::target_configs::openbsd::OpenBSDTarget;
-use reqwest;
 use std::fmt::Display;
 use std::sync::{mpsc, Arc};
-use std::time::Duration;
-use tokio::runtime::Runtime;
 use url::Url;
 
 impl LogFormatter for OpenBSDTarget {
@@ -27,17 +24,7 @@ impl FetchMirrors for OpenBSDTarget {
     ) -> Result<Vec<Mirror>, AppError> {
         let url = "https://ftp.openbsd.org/pub/OpenBSD/ftplist";
 
-        let output = Runtime::new().unwrap().block_on(async {
-            Ok::<_, AppError>(
-                reqwest::Client::new()
-                    .get(url)
-                    .timeout(Duration::from_millis(self.fetch_mirrors_timeout))
-                    .send()
-                    .await?
-                    .text_with_charset("utf-8")
-                    .await?,
-            )
-        })?;
+        let output = fetch_text(url, self.fetch_mirrors_timeout)?;
 
         let urls = output
             .lines()